fix(ios): OTS score Int + drop unreachable severity cases
Backend OTS emits integer score (0|1|2|3) and severity ∈ {low, moderate,
high}. Model the score as Int and drop dead severityTint branches that
mixed OTS and illness vocabularies.

+++ apps/ios/Sources/DashboardSummary.swift
@@ -57,7 +57,7 @@ struct DashboardSummary: Decodable {
         let apnea: Apnea?
 
         struct OTS: Decodable {
-            let score: Double
+            let score: Int
             let severity: String
         }
